implemented poisson probability

+++ AnalyseApp/Extensions/PoissonExtension.cs
@@ -1,3 +1,6 @@
+using Accord.Statistics.Distributions.Univariate;
+using MathNet.Numerics.Distributions;
+
 namespace AnalyseApp.Extensions;
 
 public static class PoissonExtensions
@@ -8,7 +11,7 @@ public static class PoissonExtensions
     /// </summary>
     /// <param name="lambda"></param>
     /// <returns></returns>
-    public static double PoissonProbability(this float lambda, int k)
+    public static double PoissonProbability(this double lambda, int k)
     {
         return Math.Pow(lambda, k) * Math.Exp(-lambda) / Factorial(k);
     }
